// frontend/src/client/types.gen.ts
export type RoleRead = {
    /**
     * Id
     */
    id: string;
    /**
     * Name
     */
    name: string;
    /**
     * Description
     */
    description: string;
    /**
     * Permission Tags
     */
    permission_tags?: Array<string>;
    /**
     * Created At
     */
    created_at: string;
};
// ...
export type UserRead = {
    /**
     * Id
     */
    id: string;
    /**
     * Username
     */
    username: string;
    /**
     * Display Name
     */
    display_name: string;
    /**
     * Status
     */
    status: 'active' | 'disabled';
    /**
     * Roles
     */
    roles?: Array<RoleRead>;
    /**
     * Created At
     */
    created_at: string;
    /**
     * Updated At
     */
    updated_at: string;
};